Remove the KimYi and KimYiPlus recoverers.
Since we now know that both are broken, the effort involved in continually
maintaining them seems pointless.

Since the MF recoverer used some parts of KimYi's code, that means that some
code has to be moved to MF.
