+++ src/lib/mod.rs
@@ -47,6 +47,4 @@ mod astar;
 mod corchuelo;
 pub mod parser;
 pub use parser::{ParseRepair, RecoveryKind};
-mod kimyi;
-mod kimyi_plus;
 mod mf;
